Work on latex->texinfo conversion.
I don't like texinfo. I really don't. For one thing, there is nothing
compelling about it; I find LaTeX to produce better output as it is.
However, automake is pretty bullheaded about using it, and until I
can master the arcane spells of m4, I need a way to generate texi
without actually having to write it. To this end, I am working on
a script and some templates to generate the texinfo file automatically.