Add rmdirs tests; add test cleanup for makedirs/rmdirs.
I wrestled with the decision to use system(3) for the cleanup;
however, because this is code to test the functionality of the
library, it is not therefore unreasonable to assume that the
makedirs and rmdirs functions do not work. Without these calls,
the test relies on manual intervention in between runs which is
not conducive to the development workflow.